vault/pyproject.toml
[project]
name = "step-vault" # wheel 名称（pip install 用）
version = "0.2.5"
description = "AI-first multimodal data storage and management library with transparent file structure"
readme = "README.md"
authors = [{ name = "wangrui", email = "wangrui@stepfun.com" }]
keywords = [
    "multimodal",
    "data-storage",
    "ai",
    "lance",
    "duckdb",
    "machine-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.2",
    "filelock>=3.19.1",
    "fire>=0.7.1",
    "fsspec>=2025.7.0",
    "loguru>=0.7.3",
    "megfile>=4.2.3",
    "opencv-python-headless>=4.12.0.88",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pyarrow>=21.0.0",
    "pylance>=0.33.0",
    "toml>=0.10.2",
    "webdataset==0.2.111",
    "xxhash>=3.5.0",
]

[project.optional-dependencies]
apps = ["gradio>=5.44.1"]
dev = [
    "ipdb>=0.13.13; python_version >= '3.11'",
    "ipython>=9.4.0; python_version >= '3.11'",
    "pytest>=8.4.1; python_version >= '3.11'",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.uv]
# index-url = "http://mirrors.i.basemind.com/pypi/simple/"
[tool.uv.pip]
index-url = "http://mirrors.i.basemind.com/pypi/simple/"
[tool.hatch.build.targets.wheel]
# 告诉 hatchling 我们的包代码在 'src/vault' 目录下。
# hatchling 会将这个目录作为名为 'vault' 的顶级包进行打包。
# 这就实现了 `pip install step-vault` 后，可以 `import vault`。
packages = ["src/vault"]
# 同时，告诉 hatchling 将 'src/vault' 目录下的 .so 和 .sql 文件也包含进去
include = ["/src/vault/**/*.so", "/src/vault/**/*.sql"]
